pyproject.toml
[project]
name = "dyxi-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "boto3>=1.42.43",
    "fastapi>=0.128.1",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "imagekitio>=5.2.0",
    "mysql-connector-python>=9.5.0",
    "passlib>=1.7.4",
    "pygame>=2.6.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
